requirements.txt
flash_attn==2.7.0.post2
matplotlib==3.7.5
numpy==1.23.5
pandas==1.5.3
PyYAML==6.0
scikit_learn==1.3.2
thop==0.1.1.post2209072238
torch==2.4.1
tqdm==4.67.1
transformers==4.46.3
